Moved wrapper auth file and re-did the referencing
