+++ src/auth/Welcome.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import styles from "./Styles.module.css";
+import styles from "../pages/Styles.module.css";
 
 function Welcome() {
     return (
